==> 1-scrape-portal.ipynb <==
import requests
import re
import pandas as pd
from bs4 import BeautifulSoup as bs
import json
from time import sleep

# ecs affidavit portal
url = 'http://103.183.38.66/'
res = requests.get(url)
print(res.status_code)

# get 2026 general election ID
general_election_id = 1
res = requests.get(f'http://103.183.38.66/election-settings/get-election-bytype?election_typeID={general_election_id}')
res_dict = res.json()
# print(res_dict)
elections = res_dict['election_schedules']
for election in elections:
    if '2026' in election['date_of_election']:
        print(election['electionID'])
        print(election['election_name'])

election_id = 478

# get election districts
res = requests.get(f'http://103.183.38.66/election-settings/get-election-zilla?electionID={election_id}')
res_dict = res.json()
# print(res_dict)
zillas = res_dict['zillas']
print(zillas[:5])

# get constituency id for each district
all_constituencies = []
for zilla in zillas:
    zillaID = zilla['zillaID']
    print(f'fetching district =', zilla['zilla_name'])
    res = requests.get(f'http://103.183.38.66/election/get-setting-constituency?zillaID={zillaID}&electionID={election_id}')
    res_dict = res.json()
    # print(res_dict)
    constituencies = res_dict['constituencies']
    print(f'constituency count =', len(constituencies))
    for cons in constituencies:
        all_constituencies.append((zillaID, cons['constituencyID'], cons['constituency_name']))
    sleep(2)
print(all_constituencies[:5])

dfcon = pd.DataFrame(all_constituencies, columns=['zillaID', 'constituencyID', 'constituency_name'])
dfcon.head()

dfcon.to_csv('data/constituencies.csv', index=False)

all_constituencies = pd.read_csv('data/constituencies.csv')
all_constituencies.head()

# get candidate info links
candidates = []
for _, cons in all_constituencies.iterrows():
    zillaID = cons['zillaID']
    constituencyID = cons['constituencyID']
    url = f'http://103.183.38.66/get/candidate/data?zilla_id={zillaID}&candidate_type=1&election_id={election_id}&constituency_id={constituencyID}'
    print(f'fetching constituency =', cons['constituency_name'])
    res = requests.get(url)
    soup = bs(res.text, 'html.parser')
    # print(soup.prettify())
    rows = soup.select('tr')
    print(f'candidates =', len(rows))
    for row in rows:
        cols = row.select('td')
        candidate_data = [zillaID, constituencyID]
        for col in cols:
            img = col.select_one('img')
            if img:
                src = img['src']
                candidate_data.append(src)
                continue
            a = col.select_one('a')
            if a:
                href = a['href']
                text = a['download']
                candidate_data.append(f'{text} | {href}')
                continue
            text = col.text.strip()
            if text != '':
                candidate_data.append(text)
                continue
        candidates.append(candidate_data)
    sleep(2)
# print(candidates)

dfcand = pd.DataFrame(candidates)
dfcand.head()

dfcand.shape

dfcand.tail()

dfcand.columns = ['zillaID', 'constituencyID', 'name', 'img', 'party', 'affidavit', 'election_fund', 'tax_return']

dfcand.to_csv('data/candidates.csv', index=False)
