==> 6-finalize-data.ipynb <==
import pandas as pd

dfcan = pd.read_csv('data/candidates.csv')
dfcan.head()

dfp = dfcan.party.value_counts().reset_index()
dfp.columns = ['party', 'counts']
dfp.head()

dfp.to_csv('data/party.csv', index=False)

dfp = pd.read_csv('data/party-en.csv')
dfp.head()

dfclean = pd.read_csv('data/data-clean-alternative.csv')
dfclean.head()

dfclean.reset_index()
dfclean.head()

dfclean = dfclean.rename(columns={'Unnamed: 0': 'filename'})
dfclean.head()

dfmp = pd.merge(dfcan, dfp, how='left', on='party')
dfmp.head()


def get_filename(row):
    return f"{row['zillaID']},{row['constituencyID']},{row['name']}.txt"

dfmp['filename'] = dfmp.apply(get_filename, axis=1)
dfmp.head()



dfmc = pd.merge(dfclean, dfmp, how='left', on='filename')
dfmc.head()

dfmc.columns

cols_to_keep = [
    'filename', 'name_bn', 'zillaID_x', 'constituencyID_x', 'tin', 'father',
    'mother', 'nid', 'current_addr', 'permanent_addr', 'shown', 'paid',
    'wealth', 'text_name', 'cert_name', 'slip_name', 'party_en', 'final', 'symbol'
]

dfmc = dfmc[cols_to_keep]
dfmc.head()

dfmc = dfmc.rename(columns={
    'zillaID_x': 'zillaID', 
    'constituencyID_x': 'constituencyID',
    'party_en': 'party',
    'shown': 'income',
    'paid': 'tax'
})
dfmc.head()

dfmc.to_csv('data/data-clean-alternative-final.csv', index=False)
